move api to utils
